requirements.txt
pandas>=1.3.0
scikit-learn>=1.0.0
numpy>=1.20.0
nltk>=3.6.0
imbalanced-learn>=0.8.0
joblib>=1.0.0
kagglehub>=0.1.0
tqdm>=4.50.0
python-dotenv>=0.19.0
torch>=1.9.0
transformers>=4.10.0
tweepy>=4.0.0
streamlit>=1.10.0